+++ src/Akeneo/Tool/Bundle/ApiBundle/Doctrine/ORM/Repository/AttributeRepository.php
@@ -3,7 +3,7 @@
 namespace Akeneo\Tool\Bundle\ApiBundle\Doctrine\ORM\Repository;
 
 use Akeneo\Pim\Structure\Component\Repository\AttributeRepositoryInterface as CatalogAttributeRepositoryInterface;
-use Akeneo\Tool\Component\Api\Repository\AttributeRepositoryInterface;
+use Akeneo\Pim\Structure\Component\Repository\ExternalApi\AttributeRepositoryInterface;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\UnexpectedResultException;
